+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Pubchem tab re-design
 - Adding more details when a note is added during formula making
 - Prevent dilutant purity to exceed 100% when composing a formula
+- Restructure ingredient overview page
 
 ### Version 8.4
 - Auto redirect to the new ingredient when is renamed